requirements.txt
livekit-agents[tavus]
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-deepgram
livekit-plugins-cartesia
livekit-plugins-groq
livekit-plugins-noise-cancellation
mem0ai
langchain_community
requests
python-dotenv
pymongo
sentence-transformers